admin/requirements.txt
asgiref==3.7.2
Django==4.1
djangorestframework==3.13.1
django-environ==0.8.1
sqlparse==0.4.4
typing_extensions==4.8.0
psycopg2==2.9.3
django-cors-headers==3.13.0
drf-spectacular==0.23.1
flake8==6.0.0
black==23.1.0
pika==1.1.0